We now sync on a background job (#137)
Major refactor for syncing the library.  I use celery and postgresql which will lead to a much more robust solution.  Syncing is now done manually via the 'sync' button instead of automatically.  Now that I've refactored quite a bit, it's probably time to step through the base cases and fix up things that don't quite work.  Then on to duplicates.

* Migrated from SQLite to PostgreSQL
* Added Celery (with a redis backend)
* PostgreSQL and Celery are now going to be prerequisites
* Changed up the server sync to be in one job since going through server / library / tracks requires nested loops